perf(mem): forced-phot loads frames as float32 (halve per-frame RAM)
Step-7 _load_image returned float64 (488 MB per 61 MP frame); float32 halves it
to 244 MB with aperture sums exact to ~1e-7 (far below the noise floor). With the
worker caps in parameters(.example).toml this drops peak Step-7 RAM from ~3 GB to
~0.5 GB, fixing the 16 GB-laptop freeze during reprocessing.

+++ apex/analysis/forced_photometry.py
@@ -550,7 +550,10 @@ def run_forced_photometry(
             with fits.open(path) as hdul:
                 for hdu in hdul:
                     if hdu.data is not None and hdu.data.ndim == 2:
-                        return hdu.data.astype(float)
+                        # float32, not float64: halves per-frame memory (a 61 MP
+                        # frame is 244 MB vs 488 MB). Aperture sums stay exact to
+                        # ~1e-7 relative — far below the photometric noise floor.
+                        return hdu.data.astype(np.float32)
         except Exception:
             pass
         return None
